// src/app/models/review.ts
export type UpdateReviewDto = {
  reviewId: string;
  ratingText: string;
  userRating: number;
};

export type CreateReviewDto = {
  hotelRoomId: string;
  ratingText: string;
  userRating: number;
  userId: string;
};

export type Review = {
  text: string;
  user: { name: string };
  userRating: number;
  _createdAt: Date;
  _id: string;
};